Rename `PreHeader` into `HeaderAccount`
